pyproject.toml
[tool.poetry]
name = "trainz-dl"
version = "0.1.0"
description = ""
authors = ["David Sn <divad.nnamtdeis@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
pydantic = "^2.6.2"
pydantic-settings = "^2.2.1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}
uvicorn = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
